+++ internal/sysutil/os.go
@@ -24,8 +24,3 @@ func HaveAdminRights() (bool, error) {
 func SendProcessSignal(pid int, sig syscall.Signal) error {
 	return sendProcessSignal(pid, sig)
 }
-
-// ConfigureSyslog reroutes standard logger output to syslog.
-func ConfigureSyslog(serviceName string) error {
-	return configureSyslog(serviceName)
-}