+++ internal/sysutil/sysutil_test.go
@@ -1,0 +1,11 @@
+package sysutil
+
+import (
+	"testing"
+
+	"github.com/AdguardTeam/AdGuardHome/internal/testutil"
+)
+
+func TestMain(m *testing.M) {
+	testutil.DiscardLogOutput(m)
+}